Dynamic CSS Styling with Class Bindings
Dynamically turn on or off any class from a CSS template based on conditions (true or false) in any component.

In this instance, it highlights the selected user—the one that was clicked. This allows us not only to view that user's tasks but also to see the same user highlighted in the user list.