Content Projection and the Use of ng-content
The goal of this lesson is to demonstrate how a single component, such as card.component, can serve as a wrapper for other components (like user.component and task.component) using ng-content. The ng-content directive acts as a placeholder inside the wrapper component (card), where the content of the child components (user and task) will be projected. At the end, the styles and content of both components are combined.