Services & Dependency Injection
A service in Angular is a class with methods that provide specific functionality.
and allow us to encapsulate and reuse business logic, making the code modular and easier to maintain.

When a service is injected into a component, Angular ensures that only one instance of that service is used throughout the app (unless specified otherwise), enabling shared state across components.

Services provide a powerful and flexible mechanism to manage business logic, shared state, and external interactions (like API calls) in a modular way. They play a key role in keeping Angular applications clean, organized, and scalable.